1
00

+++ DX11Starter/Files/Graphic/Buffer/ClusterItem.h
@@ -3,15 +3,15 @@
 #include <list>
 #include <vector>
 #include <d3d11.h>
-#include <Graphic\Vertex.h>
+#include <Graphic\Mesh\Vertex.h>
 #include <SimpleMath.h>
 #include <iostream>
 namespace NGraphic {
 	namespace NBuffer {
 		struct ClusterItem {
-			unsigned short light;
-			unsigned short decal;
-			unsigned short probe;
+			byte light;
+			byte decal;
+			byte probe;
